Solve scoping and visibility issue introduced by user agent change.
