--- Sources/Foundation/Extensions/Foundation.URLRequest.swift
@@ -1,42 +1,0 @@
-//
-//  Foundation.URLRequest.swift
-//  RoverFoundation
-//
-//  Created by Andrew Clunis on 2019-06-04.
-//  Copyright © 2019 Rover Labs Inc. All rights reserved.
-//
-
-import Foundation
-import UIKit
-
-extension URLRequest {
-    /// Replace the standard CFNetwork-created User Agent with a more useful one.
-    mutating func setRoverUserAgent() {
-        // we want to add the Rover SDK version to the User Agent. Sadly, we cannot get our hands on the default user agent sent by CFNetwork, so we have to start from scratch, including reproducing the fields in the stock user agent:
-        
-        // AppName/$version. (a value exists for this by default, but this version is a bit better than the stock version of this value, which has the app build number rather than version)
-        let appBundleDict = Bundle.main.infoDictionary
-        let appName = appBundleDict?["CFBundleName"] as? String ?? "unknown"
-        let appVersion = appBundleDict?["CFBundleShortVersionString"] as? String ?? "unknown"
-        let appDescriptor = appName + "/" + appVersion
-        
-        // CFNetwork/$version (reproducing it; it exists in the standard iOS user agent, and perhaps some ops tooling will benefit from it being stock)
-        let cfNetworkInfo = Bundle(identifier: "com.apple.CFNetwork")?.infoDictionary
-        let cfNetworkVersion = cfNetworkInfo?["CFBundleShortVersionString"] as? String ?? "unknown"
-        let cfNetworkDescriptor = "CFNetwork/\(cfNetworkVersion)"
-        
-        // Darwin/$version (reproducing it; it exists in the standard iOS user agent, and perhaps some ops tooling will benefit from it being stock)
-        var sysinfo = utsname()
-        uname(&sysinfo)
-        let dv = String(bytes: Data(bytes: &sysinfo.release, count: Int(_SYS_NAMELEN)), encoding: .ascii)?.trimmingCharacters(in: .controlCharacters) ?? "unknown"
-        let darwinDescriptor = "Darwin/\(dv)"
-        
-        // iOS/$ios-version
-        let osDescriptor = "iOS/" + UIDevice.current.systemVersion
-        
-        // RoverSDK/$rover-sdk-version
-        let roverBundle = Bundle(identifier: "io.rover.RoverFoundation") ?? Bundle(identifier: "org.cocoapods.RoverCampaigns")
-        let roverVersion = roverBundle?.infoDictionary?["CFBundleShortVersionString"] ?? "unknown"
-        self.setValue("\(appDescriptor) \(cfNetworkDescriptor) \(darwinDescriptor) \(osDescriptor) RoverCampaignsSDK/\(roverVersion)", forHTTPHeaderField: "User-Agent")
-    }
-}
